pyproject.toml
[project]
name = "legal-lingo"
version = "0.1.0"
description = "A Chatbot that helps students learn effectively, enables parents to monitor their child's progress, and allows teachers to track class performance"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "ruff>=0.11.12",
    "sqlalchemy>=2.0.41",
]
